+++ src/main/java/ru/psu/fvt/moipevm/taskmanager/services/TaskServiceImpl.java
@@ -2,7 +2,6 @@ package ru.psu.fvt.moipevm.taskmanager.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.psu.fvt.moipevm.taskmanager.TaskManagerApplication;
 import ru.psu.fvt.moipevm.taskmanager.exceptions.DeleteException;
 import ru.psu.fvt.moipevm.taskmanager.model.Task;
 import ru.psu.fvt.moipevm.taskmanager.repositories.TaskRepository;
@@ -10,7 +9,7 @@ import ru.psu.fvt.moipevm.taskmanager.repositories.TaskRepository;
 import java.util.List;
 
 @Service
-public class TaskServiceImpl implements BaseService<Task> {
+public class TaskServiceImpl implements ITaskService<Task> {
     private final TaskRepository taskRepository;
 
     @Autowired
@@ -41,10 +40,11 @@ public class TaskServiceImpl implements BaseService<Task> {
     }
 
     @Override
-    public void delete(int id) throws DeleteException {
-        if (!taskRepository.existsById(id)) {
-            throw new DeleteException("Task don't created");
-        }
+    public void delete(int id, int userId) throws DeleteException {
+        if (!taskRepository.existsTaskByUserId(userId)) throw new DeleteException("no access to task");
+
+        if (!taskRepository.existsById(id)) throw new DeleteException("Task don't created");
+
         taskRepository.deleteById(id);
     }
 }
